Some size mismatches not picked up by the 32-bit compiler
git-svn-id: https://svn.code.sf.net/p/oorexx/code-0/main/trunk@5380 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ extensions/platform/windows/oodialog/ooDialog.cpp
@@ -411,10 +411,10 @@ done_out:
     return adm;
 }
 
-HBRUSH searchForBrush(DIALOGADMIN *dlgAdm, uint32_t *index, uint32_t id)
+HBRUSH searchForBrush(DIALOGADMIN *dlgAdm, size_t *index, uint32_t id)
 {
     HBRUSH hBrush = NULL;
-    uint32_t i = 0;
+    size_t i = 0;
 
     if ( dlgAdm != NULL && dlgAdm->ColorTab != NULL )
     {
